Created Android App launch profiling post
This post outlines the various methods I figured out to profile
application launch latencies in Android.

This has one really useful method of measuring memory saturation time of
the application and using it to determine the launch time.

This might be confidential so be careful and check before sharing this.

Signed-off-by: Prathu Baronia <prathu.baronia@oneplus.com>